Change componentPreprocessor into a class
The includeFile process depends on heavily on configuration parameters
being passed repeatedly throughout its functions (in
componentPreprocessor).

Let's make componentPreprocessor a class instead, passing these
configurations as instance properties.
